+++ mijia/src/bluetooth/mod.rs
@@ -1,5 +1,8 @@
-pub use crate::bluetooth_event::BluetoothEvent;
-use crate::introspect::Node;
+mod bluetooth_event;
+mod introspect;
+
+pub use self::bluetooth_event::BluetoothEvent;
+use self::introspect::Node;
 use bluez_generated::{
     OrgBluezAdapter1, OrgBluezDevice1, OrgBluezGattCharacteristic1, OrgBluezGattService1,
 };
